# EmployeeManager/WebAPI/WebAPI/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "WebAPI, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/api/Employees": {
      "get": {
        "tags": [
          "Employees"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Employee"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Employees"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/EmployeeCreateDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Employees/{id}": {
      "get": {
        "tags": [
          "Employees"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Employee"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Employees"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Employee"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Employees"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/JobCategories": {
      "get": {
        "tags": [
          "JobCategories"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/JobCategory"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "JobCategories"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/JobCategory"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/JobCategory"
            }
          }
        }
      }
    },
    "/api/JobCategories/{id}": {
      "get": {
        "tags": [
          "JobCategories"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/JobCategory"
            }
          }
        }
      },
      "put": {
        "tags": [
          "JobCategories"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/JobCategory"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "JobCategories"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
    "Address": {
      "required": [
        "city",
        "cityId",
        "country",
        "countryId",
        "isActive",
        "street",
        "zipCode"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "street": {
          "maxLength": 250,
          "minLength": 3,
          "type": "string"
        },
        "zipCode": {
          "maxLength": 5,
          "minLength": 0,
          "type": "string"
        },
        "city": {
          "$ref": "#/definitions/City"
        },
        "cityId": {
          "format": "int32",
          "type": "integer"
        },
        "country": {
          "$ref": "#/definitions/Country"
        },
        "countryId": {
          "format": "int32",
          "type": "integer"
        },
        "isActive": {
          "type": "boolean"
        }
      }
    },
    "AddressCreateDTO": {
      "required": [
        "cityId",
        "countryId",
        "street",
        "zipCode"
      ],
      "type": "object",
      "properties": {
        "street": {
          "maxLength": 250,
          "minLength": 3,
          "type": "string"
        },
        "zipCode": {
          "maxLength": 5,
          "minLength": 0,
          "type": "string"
        },
        "cityId": {
          "format": "int32",
          "type": "integer"
        },
        "countryId": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "City": {
      "required": [
        "country",
        "countryId",
        "name"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "maxLength": 100,
          "minLength": 1,
          "type": "string"
        },
        "country": {
          "$ref": "#/definitions/Country"
        },
        "countryId": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "Country": {
      "required": [
        "name"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "maxLength": 50,
          "minLength": 2,
          "type": "string"
        }
      }
    },
    "EGender": {
      "format": "int32",
      "enum": [
        0,
        1,
        2
      ],
      "type": "integer"
    },
    "Employee": {
      "required": [
        "address",
        "addressId",
        "birthDate",
        "country",
        "countryId",
        "email",
        "firstName",
        "gender",
        "joinedDate",
        "lastName",
        "phoneNumber"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "firstName": {
          "maxLength": 250,
          "minLength": 1,
          "type": "string"
        },
        "middleName": {
          "maxLength": 250,
          "minLength": 1,
          "type": "string"
        },
        "lastName": {
          "maxLength": 250,
          "minLength": 1,
          "type": "string"
        },
        "birthDate": {
          "format": "date-time",
          "type": "string"
        },
        "gender": {
          "$ref": "#/definitions/EGender"
        },
        "address": {
          "$ref": "#/definitions/Address"
        },
        "addressId": {
          "format": "int32",
          "type": "integer"
        },
        "country": {
          "$ref": "#/definitions/Country"
        },
        "countryId": {
          "format": "int32",
          "type": "integer"
        },
        "email": {
          "maxLength": 250,
          "minLength": 3,
          "type": "string"
        },
        "phoneNumber": {
          "maxLength": 250,
          "minLength": 3,
          "type": "string"
        },
        "joinedDate": {
          "format": "date-time",
          "type": "string"
        },
        "exitedDay": {
          "format": "date-time",
          "type": "string"
        },
        "superior": {
          "$ref": "#/definitions/Employee"
        },
        "superiorId": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "EmployeeCreateDTO": {
      "required": [
        "addressCreateDTO",
        "birthDate",
        "email",
        "firstName",
        "gender",
        "jobCategoryId",
        "lastName",
        "phoneNumber",
        "salary"
      ],
      "type": "object",
      "properties": {
        "firstName": {
          "maxLength": 250,
          "minLength": 1,
          "type": "string"
        },
        "middleName": {
          "type": "string"
        },
        "lastName": {
          "maxLength": 250,
          "minLength": 1,
          "type": "string"
        },
        "birthDate": {
          "format": "date-time",
          "type": "string"
        },
        "gender": {
          "$ref": "#/definitions/EGender"
        },
        "addressCreateDTO": {
          "$ref": "#/definitions/AddressCreateDTO"
        },
        "email": {
          "maxLength": 250,
          "minLength": 3,
          "type": "string"
        },
        "phoneNumber": {
          "maxLength": 250,
          "minLength": 3,
          "type": "string"
        },
        "superiorId": {
          "format": "int32",
          "type": "integer"
        },
        "salary": {
          "format": "double",
          "minimum": 1,
          "type": "number"
        },
        "jobCategoryId": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "JobCategory": {
      "required": [
        "title"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "title": {
          "maxLength": 150,
          "minLength": 3,
          "type": "string"
        }
      }
    }
  }
}